pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["map_builder_service*"]

[project]
name = "map-builder-service"
description = ""
version = "0.0.1"
readme = "README.md"
keywords = []
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
]
dependencies = [
  "dvc>=2.47.1",
  "dvc-s3",
  "fire",
  "tqdm",
  "lightning[extra]",
  "torch",
  "torchvision",
  "torchmetrics",
  "click",
  "pillow",
  "scikit-learn",
  "opencv-python",
]

[project.optional-dependencies]
dev = [
]
lint = [
     "flake8",
     "black",
     "ruff",
     "mypy",
     "types-PyYAML",
     "types-pkg_resources",
     "types-requests",
]
test = [
     "pytest",
     "pytest-sugar",
     "pytest-datadir-ng",
     "pytest-repeat",
     "pytest-benchmark",
]

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ignore=setup.py"
testpaths = [
  "tests",
]

[tool.mypy]
ignore_missing_imports = true

[tool.ruff]
select = [
  "E", # pycodestyle
  "F", # Pyflakes
  "I", # isort
  "SLF", # flake8-self
  "T20", # flake8-print
  "PIE", # flake8-pie
  "TRY", # tryceratops
  "C90", # mccabe
]
line-length = 88

[tool.black]
line-length = 88
target-version = ['py310']
